requirements.txt
opencv-python
langchain
langchain-openai
langchain_google_genai
langchain-community
python-dotenv
pyaudio
soundfile
SpeechRecognition
pillow==10.0.0
pyautogui==0.9.54
git+https://github.com/openai/whisper.git
